Two-stage in-flight system
Remember all in-flight requests, with their ID, Method, and the
timestamp at which they were dispatched.

After shutdown is 'initiated', check, any time a request completes,
if by chance there are no in-flight requests left.

If that's the case, then 'complete' the shutdown (as far as the router
is concerned) by requesting the HTTP server to be shut down. They do
as soon as all the connections are closed.

Finally, butler exits with exit code 0.