Don't prefer unpacked when installing.
Heeeeeere's the tea:
- Linux folks wanted to have butler's backend generate .tar.gz files
- I said "oh, ah, maybe later"
- Said folks didn't want to wait
- They started pushing .tar.gz files with butler
- Which worked great for web, since they'd get *the only file in the
container* as unpacked, ie. the .tar.gz. file - but also completely
defeats patching.
- It didn't work as great for app users, which, well, saw only
the .tar.gz file in their caves.
- So I started preferring "unpacked" build files on install, so that
butler would detect that they're a supported archive format, and
extract-them-while-downloading, as we can do with .tar.gz files
- This made some folks happy but it also broke prereqs for a while,
since some prereqs are a single (windows installer) file, and
the install logic is used to download prereqs to some folder, and so
butler just tried to install them before it was time to install them
(usually as admin).
- *Additionally*, this particular set up (single .tar.gz file pushed)
broke reinstall (heal), downgrade (heal), upgrade (patching, which would
fail and fall back to heal).
- tl;dr I'm reversing my decision - folks pushing .tar.gz will have the
app install that .tar.gz in caves for now. This workaround just isn't
worth it for now.