Added a --resume-file argument to butler extract
I added the ability to interrupt downloads part way through using a lot of existing features in the stack. itchio/hush provides an intervalSaveConsumer that can store the results of a partial extraction. And itchio/savior can use this as part of the normal extraction.

My code conditionally creates a saveConsumer if the --resume-file argument is provided and gives it to the extractor instead of a nil checkpoint to ex.Resume()

I haven't worked with golang before, but I compiled a local version of butler that seems to work.